Move array bracket pairs from Type to PrimitiveType and ClassOrInterfaceType
