+++ javaparser-core/src/main/java/com/github/javaparser/ast/type/ClassOrInterfaceType.java
@@ -24,8 +24,10 @@ package com.github.javaparser.ast.type;
 import java.util.List;
 
 import com.github.javaparser.Range;
+import com.github.javaparser.ast.ArrayBracketPair;
 import com.github.javaparser.ast.TypeArguments;
 import com.github.javaparser.ast.nodeTypes.NodeWithAnnotations;
+import com.github.javaparser.ast.nodeTypes.NodeWithArrayBrackets;
 import com.github.javaparser.ast.nodeTypes.NodeWithName;
 import com.github.javaparser.ast.visitor.GenericVisitor;
 import com.github.javaparser.ast.visitor.VoidVisitor;
@@ -33,13 +35,18 @@ import com.github.javaparser.ast.visitor.VoidVisitor;
 /**
  * @author Julio Vilmar Gesser
  */
-public final class ClassOrInterfaceType extends ReferenceType<ClassOrInterfaceType> implements NodeWithName<ClassOrInterfaceType>, NodeWithAnnotations<ClassOrInterfaceType> {
+public final class ClassOrInterfaceType extends ReferenceType<ClassOrInterfaceType> implements
+        NodeWithName<ClassOrInterfaceType>,
+        NodeWithAnnotations<ClassOrInterfaceType>,
+        NodeWithArrayBrackets<ClassOrInterfaceType> {
 
     private ClassOrInterfaceType scope;
 
     private String name;
 
     private TypeArguments typeArguments = TypeArguments.EMPTY;
+
+    private List<ArrayBracketPair> arrayBracketPairs;
 
     public ClassOrInterfaceType() {
     }
@@ -122,5 +129,20 @@ public final class ClassOrInterfaceType extends ReferenceType<ClassOrInterfaceTy
     public void setTypeArguments(TypeArguments typeArguments) {
         this.typeArguments = typeArguments;
         setAsParentNodeOf(this.typeArguments.getTypeArguments());
+    }
+
+    /**
+     * @return the array brackets directly after a type.
+     */
+    @Override
+    public List<ArrayBracketPair> getArrayBracketPairs() {
+        return arrayBracketPairs;
+    }
+
+    @Override
+    public ClassOrInterfaceType setArrayBracketPairs(List<ArrayBracketPair> arrayBracketPairs) {
+        this.arrayBracketPairs = arrayBracketPairs;
+        setAsParentNodeOf(arrayBracketPairs);
+        return this;
     }
 }
