+++ javaparser-core/src/main/java/com/github/javaparser/ast/type/PrimitiveType.java
@@ -25,15 +25,18 @@ import java.util.HashMap;
 import java.util.List;
 
 import com.github.javaparser.Range;
-import com.github.javaparser.ast.expr.AnnotationExpr;
+import com.github.javaparser.ast.ArrayBracketPair;
 import com.github.javaparser.ast.nodeTypes.NodeWithAnnotations;
+import com.github.javaparser.ast.nodeTypes.NodeWithArrayBrackets;
 import com.github.javaparser.ast.visitor.GenericVisitor;
 import com.github.javaparser.ast.visitor.VoidVisitor;
 
 /**
  * @author Julio Vilmar Gesser
  */
-public final class PrimitiveType extends Type<PrimitiveType> implements NodeWithAnnotations<PrimitiveType> {
+public final class PrimitiveType extends Type<PrimitiveType> implements
+		NodeWithAnnotations<PrimitiveType>,
+		NodeWithArrayBrackets<PrimitiveType> {
 
 	public static final PrimitiveType BYTE_TYPE = new PrimitiveType(Primitive.Byte);
 
@@ -81,6 +84,8 @@ public final class PrimitiveType extends Type<PrimitiveType> implements NodeWith
 
 	private Primitive type;
 
+	private List<ArrayBracketPair> arrayBracketPairs;
+
 	public PrimitiveType() {
 	}
 
@@ -113,4 +118,18 @@ public final class PrimitiveType extends Type<PrimitiveType> implements NodeWith
 		this.type = type;
 	}
 
+	/**
+	 * @return the array brackets directly after a type.
+	 */
+	@Override
+	public List<ArrayBracketPair> getArrayBracketPairs() {
+		return arrayBracketPairs;
+	}
+
+	@Override
+	public PrimitiveType setArrayBracketPairs(List<ArrayBracketPair> arrayBracketPairs) {
+		this.arrayBracketPairs = arrayBracketPairs;
+		setAsParentNodeOf(arrayBracketPairs);
+		return this;
+	}
 }
