+++ javaparser-core/src/main/java/com/github/javaparser/ast/type/TypeParameter.java
@@ -26,7 +26,6 @@ import static com.github.javaparser.utils.Utils.ensureNotNull;
 import java.util.List;
 
 import com.github.javaparser.Range;
-import com.github.javaparser.ast.Node;
 import com.github.javaparser.ast.expr.AnnotationExpr;
 import com.github.javaparser.ast.nodeTypes.NodeWithName;
 import com.github.javaparser.ast.visitor.GenericVisitor;
